requirements.txt
streamlit
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0   
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics
Pillow
numpy
opencv-python
matplotlib
pandas
scikit-learn
scipy
